resolve conflicts by restructuring folders as they were before
