+++ README.md
@@ -1,0 +1,5 @@
+230705-2 - version 0.1.2-stable (branch) stable version with internal dredd execution engine plus engine reset button
+
+230705-1 - version 0.1.1-stable (branch) stable version with internal dredd execution engine
+
+230705-0 - version 0.1 This branch is deployed to latest.trypennant.com as pennantmvp2.fly.dev\*\*\*\*
