+++ client/fly.toml
@@ -1,0 +1,15 @@
+# fly.toml app configuration file generated for dreddpennant on 2023-07-05T22:48:29-07:00
+#
+# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
+#
+
+app = "dreddpennant"
+primary_region = "sea"
+
+[http_service]
+  internal_port = 8080
+  force_https = true
+  auto_stop_machines = true
+  auto_start_machines = true
+  min_machines_running = 0
+  processes = ["app"]
