+++ client/src/utils/yPrettyPrint.js
@@ -1,0 +1,3 @@
+export const yPrettyPrint = (ydoc, msg = '') => {
+  console.log('\n\n==> ' + msg + ': \n' + JSON.stringify(ydoc.toJSON(), null, 4) + '\n\n');
+};
