+++ server/fly.toml
@@ -1,0 +1,15 @@
+# fly.toml app configuration file generated for s3pocusui on 2023-07-12T18:06:52-07:00
+#
+# See https://fly.io/docs/reference/configuration/ for information about how to use this file.
+#
+
+app = "s3pocusui"
+primary_region = "sea"
+
+[http_service]
+  internal_port = 3000
+  force_https = true
+  auto_stop_machines = true
+  auto_start_machines = true
+  min_machines_running = 0
+  processes = ["app"]
