+++ server/package.json
@@ -3,11 +3,14 @@
   "version": "1.0.0",
   "description": "",
   "type": "module",
-  "main": "app.js",
+  "main": "src/index.js",
   "scripts": {
-    "test": "echo \"Error: no test specified\" && exit 1",
-    "dev": "nodemon app.js",
-    "start": "concurrently \"npm run dev\" \"cd ../client && npm run dev\""
+    "start": "node src/index.js",
+    "dev": "nodemon src/index.js",
+    "dev:all": "concurrently \"npm run dev\" \"cd ../client && npm run dev\"",
+    "build:ui": "cd ../client && npm run build && cp -r dist ../server",
+    "deploy:full": "npm run build:ui && git add . && git commit -m uibuild && npm run deploy",
+    "test": "echo \"Error: no test specified\" && exit 1"
   },
   "keywords": [],
   "author": "",
